Inventory System
Player now has 5 slots of inventory. Pick up an item tagged "canPickUp" to add it to the inventory. To drop/throw item, press the number associated with the inventory slot and drop/throw as before.

For future item, add a sprite to create its image for the inventory slot. As of now I only have the image for the box item.